Move status types to new file
